+++ src/sys/hw/mcu/avr/atmega328p/namespace.rs
@@ -5,7 +5,7 @@
 // 35 Register Summary:
 // https://onlinedocs.microchip.com/oxy/GUID-0EC909F9-8FB7-46B2-BF4B-05290662B5C3-en-US-12.1.1/GUID-05F454E7-6DE3-4D77-9C23-FEBE05568812.html
 
-use crate::{AvrPort, AvrUsart};
+use crate::{AvrPort, AvrTimer0, AvrUsart};
 
 #[doc = crate::_tags!(hw namespace)]
 /// ATmega328P microcontroller namespace.
@@ -34,6 +34,9 @@ impl Atmega328p {
 
     /// GPIO port D.
     pub const PORT_D: AvrPort = AvrPort::new(0x29, 0x2A, 0x2B);
+
+    /// Timer/Counter 0 peripheral.
+    pub const TIMER_0: AvrTimer0 = AvrTimer0::new(0x44, 0x45, 0x46, 0x47, 0x48, 0x6E, 0x35);
 
     /// USART 0 peripheral.
     pub const USART_0: AvrUsart = AvrUsart::new(0xC0, 0xC1, 0xC2, 0xC4, 0xC5, 0xC6);